pyproject.toml
[project]
name = "ovh-dns-updater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ovh>=1.2.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
